# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vivid Detailing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking (status, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/abandon:
    post:
      operationId: abandonBooking
      tags: [bookings]
      summary: Mark a booking as abandoned (triggers GHL webhook)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lastStep:
                  type: integer
              required: [lastStep]
      responses:
        "200":
          description: Booking abandoned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  # ─── Customers ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
          nullable: true
        vehicleId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        appointmentAt:
          type: string
          nullable: true
        totalEstimate:
          type: number
          nullable: true
        depositPaid:
          type: boolean
        notes:
          type: string
          nullable: true
        ghlContactId:
          type: string
          nullable: true
        ghlOpportunityId:
          type: string
          nullable: true
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BookingItem"
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        vehicle:
          $ref: "#/components/schemas/Vehicle"
          nullable: true
      required: [id, status, depositPaid, createdAt, items]
    BookingItem:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        itemType:
          type: string
          enum: [service, addon, quote, promo]
        itemName:
          type: string
        unitPrice:
          type: number
          nullable: true
        quantity:
          type: integer
        isQuoteBased:
          type: boolean
      required: [id, bookingId, itemType, itemName, quantity, isQuoteBased]
    CreateBookingBody:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/UpsertCustomerBody"
        vehicle:
          $ref: "#/components/schemas/CreateVehicleBody"
        serviceIds:
          type: array
          items:
            type: string
        addOnIds:
          type: array
          items:
            type: string
        promoIds:
          type: array
          items:
            type: string
        appointmentAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        totalEstimate:
          type: number
          nullable: true
      required: [customer, vehicle, serviceIds, addOnIds, promoIds]
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        appointmentAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        depositPaid:
          type: boolean
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        ghlContactId:
          type: string
          nullable: true
        ghlSyncStatus:
          type: string
          nullable: true
        ghlLastSync:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, createdAt]
    UpsertCustomerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [name, email, phone]
    Vehicle:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        type:
          type: string
          enum: [car, suv, truck, van]
        year:
          type: integer
          nullable: true
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        colour:
          type: string
          nullable: true
        licensePlate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, customerId, type, createdAt]
    CreateVehicleBody:
      type: object
      properties:
        customerId:
          type: string
          nullable: true
        type:
          type: string
          enum: [car, suv, truck, van]
        year:
          type: integer
          nullable: true
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        colour:
          type: string
          nullable: true
        licensePlate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [type]
...